Remove/fix TODOs in source code
