+++ Sources/LeafKit/LeafLexer/LeafRawTemplate.swift
@@ -1,4 +1,3 @@
-// TODO: Make `LeafSource` return this instead of `ByteBuffer` via extension
 internal struct LeafRawTemplate {
     // MARK: - Internal Only
     let name: String
