+++ Sources/LeafKit/LeafLexer/LeafToken.swift
@@ -21,8 +21,6 @@
 /// - `.stringLiteral`: Does not appear to be used anywhere?
 /// - `.whitespace`: Only generated when not at top-level, and unclear why maintaining it is useful
 ///
-/// # TODO
-/// - LeafTokens would ideally also store the range of their location in the original source template
 
 internal enum LeafToken: CustomStringConvertible, Equatable  {
     /// Holds a variable-length string of data that will be passed through with no processing
